Fix Repository definitions for Privilege and Role
Use `Integer` for ID data type

+++ src/main/java/org/launchcode/codingevents/data/PrivilegeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PrivilegeRepository extends CrudRepository<Privilege, Long> {
+public interface PrivilegeRepository extends CrudRepository<Privilege, Integer> {
 
     Privilege findByName(String name);
